Enforce no period at the end of the subject (#14)
* Enforce no period at the end of the subject

According to [this blog post](https://chris.beams.io/posts/git-commit/),
one of the golden rules for good commit messages is to not have a period
at the end of the subject. We should enforce this as part of this GH action.

* Add period restriction to README

Also link to the blog post that we got the guidelines from.